requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==1.4
pydantic
pymysql
python-dotenv
